Rename top-level E2E tests to e2e-test
